+++ apps/web/src/app/(needLogin)/mypage/like_story/_components/LikesStory/LikesStory.tsx
@@ -2,9 +2,9 @@ import { useRouter } from 'next/navigation';
 import { useEffect } from 'react';
 
 import { CommonStory, Pagination } from '@/components';
+import { usePagination } from '@/hooks/usePagination';
 
 import { useLikeStory } from '../../_hooks/useLikeStory';
-import { usePagination } from '../../_hooks/usePagination';
 
 export const LikeStory = () => {
   const router = useRouter();
