--- apps/web/src/app/(needLogin)/mypage/setting/service/_hooks/useOneInquiryHistory.ts
@@ -1,15 +1,0 @@
-import { useAtom } from 'jotai';
-
-export const useToggle = (atomCreator, id) => {
-  const [toggleState, setToggleState] = useAtom(atomCreator(id));
-
-  const handleToggle = () => {
-    setToggleState((prev: boolean) => !prev);
-  };
-
-  return {
-    toggleState,
-    setToggleState,
-    handleToggle,
-  };
-};
